Add delete/ unfollow function and play functionality

+++ src/controllers/playlist.js
@@ -4,7 +4,8 @@ const SHA256 = require('crypto-js/sha256');
 const { getUserPlaylistsSpotify } = require('../spotifyControllers');
 const PlaylistModel = require('../models/playlist');
 const UserModel = require('../models/user');
-const https = require('https');
+const { startPlayback } = require('../spotifyControllers');
+const { unfollowPlaylistSpotify } = require('../spotifyControllers');
 const { createPlaylist } = require('../spotifyControllers');
 const { addSongToPlaylist } = require('../spotifyControllers');
 const { getAudioFeaturesForTracks } = require('../spotifyControllers');
@@ -279,13 +280,29 @@ const read_helper = async (user, playlistSpotify) => {
 
 const remove = async (req, res) => {
     try {
-        // find and remove playlist
-        await PlaylistModel.findByIdAndRemove(req.params.id).exec();
+        const playlist = await PlaylistModel.findById(req.params.id).exec();
+
+        // Delete/ unfollow that playlist on spotify
+        console.log('req.userId: ', req.userId);
+        const user = await UserModel.findOne({
+            _id: req.userId,
+        }).exec();
+        const result = await unfollowPlaylistSpotify(user, playlist.spotify_id)
+
+        if (result) {
+            // find and remove playlist
+            var deletedPlaylist = await PlaylistModel.findByIdAndRemove(req.params.id).exec();
+        } else {
+            throw new Error("No result from Spotify");
+        }
 
         // return message that playlist was deleted
         return res
             .status(200)
-            .json({ message: `Playlist with id${req.params.id} was deleted` });
+            .json({
+                message: `Playlist with id${deletedPlaylist._id} was deleted`,
+                removedPlaylistId: deletedPlaylist._id,
+            });
     } catch (err) {
         console.log(err);
         return res.status(500).json({
@@ -294,6 +311,33 @@ const remove = async (req, res) => {
         });
     }
 };
+
+const play = async (req, res) => {
+    try {
+        // create the uri to play
+        const playlist = await PlaylistModel.findById(req.params.id).exec();
+        const uri = 'spotify:playlist:' + playlist.spotify_id;
+        console.log('playlist.spotify_id: ', playlist.spotify_id);
+
+        // Delete/ unfollow that playlist on spotify
+        const user = await UserModel.findOne({
+            _id: req.userId,
+        }).exec();
+        const result = await startPlayback(user, uri);
+        console.log('result: ', result);
+
+        // return message that playlist was deleted
+        return res
+            .status(200)
+            .json({ message: 'Started playback', playlistId: playlist._id });
+    } catch (err) {
+        console.log(err);
+        return res.status(500).json({
+            error: 'Internal server error',
+            message: err.message,
+        });
+    }
+}
 
 const list = async (req, res) => {
     try {
@@ -690,4 +734,5 @@ module.exports = {
     get_playlist_time,
     getAllTrackIDs,
     follow,
+    play,
 };
